requirements.txt
diffusers==0.27.2
hydra-core==1.3.2
numpy==1.23.2
omegaconf==2.2.3
Pillow==10.2.0
PyYAML==6.0.1
timm==0.4.12
torch==1.12.1+cu116
torchvision==0.13.1+cu116
tqdm==4.64.1
wandb
